+++ src/docker.py
@@ -1,11 +1,11 @@
 import logging
 import os
 import docker
+from src import settings
 
 
 class Docker:
     _DOCKER_SOCK = "/var/run/docker.sock"
-    _IMAGE_REF = "jan-di/database-backup"
 
     def __init__(self, config):
         self._config = config
@@ -17,19 +17,22 @@ class Docker:
 
         self._client = docker.from_env()
 
+        logging.debug(f"Instance ID: {config.instance_id}")
+
         containers = self._client.containers.list(
             filters={
                 "status": "running",
-                "label": f"org.opencontainers.image.ref.name={self._IMAGE_REF}",
+                "label": f"{settings.LABEL_PREFIX}instance_id={config.instance_id}",
             }
         )
 
         if len(containers) == 0:
-            raise RuntimeError("Cannot determine own container id!")
+            raise RuntimeError(
+                f"Cannot determine own container id! InstanceId: {config.instance_id}")
         elif len(containers) > 1:
             raise RuntimeError(
-                "Detected another instance of this image."
-                " Running multiple instances of the backup service is currently not supported!"
+                f"Detected another instance of this image with the same instance id. InstanceId: {config.instance_id}"
+                " To run multiple instances of the service, each container must have a unique instance id assigned."
             )
 
         self._own_container = containers[0]
@@ -42,8 +45,9 @@ class Docker:
         )
 
     def create_backup_network(self):
-        self._network = self._client.networks.create(
-            self._config.docker_network_name)
+        network_name = f"{self._config.docker_network_name}_{self._config.instance_id}"
+
+        self._network = self._client.networks.create(network_name)
         self._network.connect(self._own_container.id)
 
     def remove_backup_network(self):
@@ -52,9 +56,13 @@ class Docker:
             self._network.remove()
             self._network = None
 
+    def get_target_name(self, container):
+        return f"{self._config.docker_target_name}_{self._config.instance_id}"
+
     def connect_target(self, container):
-        self._network.connect(container, aliases=[
-                              self._config.docker_target_name])
+        target_name = self.get_target_name(container)
+
+        self._network.connect(container, aliases=[target_name])
 
     def disconnect_target(self, container):
         self._network.disconnect(container)